Move CLI commands to seperate groups
