+++ jupyter_cache/cli/commands/__init__.py
@@ -1,2 +1,4 @@
 from .cmd_cache import *  # noqa: F401,F403
+from .cmd_config import *  # noqa: F401,F403
 from .cmd_exec import *  # noqa: F401,F403
+from .cmd_stage import *  # noqa: F401,F403
