+++ jupyter_cache/cli/commands/cmd_cache.py
@@ -1,4 +1,3 @@
-from pathlib import Path
 import sys
 
 import click
@@ -7,6 +6,7 @@ import yaml
 
 from jupyter_cache.cli.commands.cmd_main import jcache
 from jupyter_cache.cli import arguments, options
+from jupyter_cache.cli.utils import shorten_path
 from jupyter_cache.cache import JupyterCacheBase
 from jupyter_cache.base import (  # noqa: F401
     CachingError,
@@ -15,33 +15,10 @@ from jupyter_cache.base import (  # noqa: F401
 )
 
 
-def shorten_path(file_path, length):
-    """Split the path into separate parts,
-    select the last 'length' elements and join them again
-    """
-    if length is None:
-        return Path(file_path)
-    return Path(*Path(file_path).parts[-length:])
-
-
-@jcache.command("clear")
-@options.CACHE_PATH
-def clear_cache(cache_path):
-    """Clear the cache completely."""
-    db = JupyterCacheBase(cache_path)
-    click.confirm("Are you sure you want to permanently clear the cache!?", abort=True)
-    db.clear_cache()
-    click.secho("Cache cleared!", fg="green")
-
-
-@jcache.command("cache-limit")
-@options.CACHE_PATH
-@click.argument("limit", metavar="CACHE_LIMIT", type=int)
-def change_cache_limit(cache_path, limit):
-    """Change the maximum number of notebooks stored in the cache."""
-    db = JupyterCacheBase(cache_path)
-    db.change_cache_limit(limit)
-    click.secho("Cache limit changed!", fg="green")
+@jcache.group("cache")
+def cmnd_cache():
+    """Commands for adding to and inspecting the cache."""
+    pass
 
 
 def format_cache_record(record, hashkeys, path_length):
@@ -57,7 +34,7 @@ def format_cache_record(record, hashkeys, path_length):
     return data
 
 
-@jcache.command("list-cached")
+@cmnd_cache.command("list")
 @options.CACHE_PATH
 @click.option("-h", "--hashkeys", is_flag=True, help="Whether to show hashkeys.")
 @options.PATH_LENGTH
@@ -79,7 +56,7 @@ def list_caches(cache_path, hashkeys, path_length):
     )
 
 
-@jcache.command("show-cached")
+@cmnd_cache.command("show")
 @options.CACHE_PATH
 @arguments.PK
 def show_cache(cache_path, pk):
@@ -101,7 +78,7 @@ def show_cache(cache_path, pk):
         click.echo(yaml.safe_dump({"Data": record.data}))
 
 
-@jcache.command("cat-artifact")
+@cmnd_cache.command("cat-artifact")
 @options.CACHE_PATH
 @arguments.PK
 @arguments.ARTIFACT_RPATH
@@ -151,7 +128,7 @@ def cache_file(db, nbpath, validate, overwrite, artifact_paths=()):
     return True
 
 
-@jcache.command("cache-nb")
+@cmnd_cache.command("add-one")
 @arguments.ARTIFACT_PATHS
 @options.NB_PATH
 @options.CACHE_PATH
@@ -165,7 +142,7 @@ def cache_nb(cache_path, artifact_paths, nbpath, validate, overwrite):
         click.secho("Success!", fg="green")
 
 
-@jcache.command("cache-nbs")
+@cmnd_cache.command("add-many")
 @arguments.NB_PATHS
 @options.CACHE_PATH
 @options.VALIDATE_NB
@@ -182,7 +159,7 @@ def cache_nbs(cache_path, nbpaths, validate, overwrite):
         click.secho("Success!", fg="green")
 
 
-@jcache.command("remove-cached")
+@cmnd_cache.command("remove")
 @arguments.PKS
 @options.CACHE_PATH
 @options.REMOVE_ALL
@@ -204,7 +181,7 @@ def remove_caches(cache_path, pks, remove_all):
     click.secho("Success!", fg="green")
 
 
-@jcache.command("diff-nb")
+@cmnd_cache.command("diff-nb")
 @arguments.PK
 @arguments.NB_PATH
 @options.CACHE_PATH
@@ -213,98 +190,3 @@ def diff_nb(cache_path, pk, nbpath):
     db = JupyterCacheBase(cache_path)
     click.echo(db.diff_nbfile_with_cache(pk, nbpath, as_str=True))
     click.secho("Success!", fg="green")
-
-
-@jcache.command("stage-nbs")
-@arguments.NB_PATHS
-@options.CACHE_PATH
-def stage_nbs(cache_path, nbpaths):
-    """Stage notebook(s) for execution."""
-    db = JupyterCacheBase(cache_path)
-    for path in nbpaths:
-        # TODO deal with errors (print all at end? or option to ignore)
-        click.echo("Staging: {}".format(path))
-        db.stage_notebook_file(path)
-    click.secho("Success!", fg="green")
-
-
-@jcache.command("stage-nb")
-@arguments.ASSET_PATHS
-@options.NB_PATH
-@options.CACHE_PATH
-def stage_nb(cache_path, nbpath, asset_paths):
-    """Cache a notebook, with possible assets."""
-    db = JupyterCacheBase(cache_path)
-    db.stage_notebook_file(nbpath, asset_paths)
-    click.secho("Success!", fg="green")
-
-
-@jcache.command("unstage-nbs")
-@arguments.NB_PATHS
-@options.CACHE_PATH
-@options.REMOVE_ALL
-def unstage_nbs(cache_path, nbpaths, remove_all):
-    """Unstage notebook(s) for execution."""
-    db = JupyterCacheBase(cache_path)
-    if remove_all:
-        nbpaths = [record.uri for record in db.list_staged_records()]
-    for path in nbpaths:
-        # TODO deal with errors (print all at end? or option to ignore)
-        click.echo("Unstaging: {}".format(path))
-        db.discard_staged_notebook(path)
-    click.secho("Success!", fg="green")
-
-
-def format_staged_record(record, cache_record, path_length, assets=True):
-    data = {
-        "ID": record.pk,
-        "URI": str(shorten_path(record.uri, path_length)),
-        "Created": record.created.isoformat(" ", "minutes"),
-    }
-    if assets:
-        data["Assets"] = len(record.assets)
-    if cache_record:
-        data["Cache ID"] = cache_record.pk
-    return data
-
-
-@jcache.command("list-staged")
-@options.CACHE_PATH
-@click.option(
-    "--compare/--no-compare",
-    default=True,
-    show_default=True,
-    help="Compare to cached notebooks (to find cache ID).",
-)
-@options.PATH_LENGTH
-def list_staged(cache_path, compare, path_length):
-    """List notebooks staged for possible execution."""
-    db = JupyterCacheBase(cache_path)
-    records = db.list_staged_records()
-    if not records:
-        click.secho("No Staged Notebooks", fg="blue")
-    rows = []
-    for record in sorted(records, key=lambda r: r.created, reverse=True):
-        cache_record = None
-        if compare:
-            cache_record = db.get_cache_record_of_staged(record.uri)
-        rows.append(format_staged_record(record, cache_record, path_length))
-    click.echo(tabulate.tabulate(rows, headers="keys"))
-
-
-@jcache.command("show-staged")
-@options.CACHE_PATH
-@arguments.PK
-def show_staged(cache_path, pk):
-    """Show details of a staged notebook."""
-    db = JupyterCacheBase(cache_path)
-    record = db.get_staged_record(pk)
-    cache_record = db.get_cache_record_of_staged(record.uri)
-    data = format_staged_record(record, cache_record, None, assets=False)
-    click.echo(yaml.safe_dump(data, sort_keys=False), nl=False)
-    if not record.assets:
-        click.echo("")
-        return
-    click.echo(f"Assets:")
-    for path in record.assets:
-        click.echo(f"- {path}")
