+++ jupyter_cache/cli/commands/cmd_exec.py
@@ -21,7 +21,7 @@ click_log.basic_config(logger)
 @options.EXEC_ENTRYPOINT
 @options.CACHE_PATH
 def execute_nbs(cache_path, entry_point):
-    """Execute outdated notebooks."""
+    """Execute staged notebooks that are outdated."""
     from jupyter_cache.executors import load_executor
 
     db = JupyterCacheBase(cache_path)
